64-bit support: cast NSInteger/NSUInteger to long/unsigned long in format functions (#249)
this should have been done already in commit 911a2a3; I now see the
error of my ways...

+++ src/archive/ArchiveViewController.m
@@ -286,7 +286,7 @@ enum DeleteAllSectionItem
   assert(editingStyle == UITableViewCellEditingStyleDelete);
   if (editingStyle != UITableViewCellEditingStyleDelete)
   {
-    DDLogError(@"%@: Unexpected editingStyle %ld", self, editingStyle);
+    DDLogError(@"%@: Unexpected editingStyle %d", self, (int)editingStyle);
     return;
   }
